Rename the operator menu row to POSStaffMenuRow
Address review nit: use "Staff" rather than "operator" for the signed-in
staff row — the view, its accessibility identifier, and the label-format
string key.
